feat: added new component
