+++ crowdin/messages.json
@@ -1011,7 +1011,6 @@
   "1375342464": "Deriv MT5 signals",
   "1376724761": "We provide online trading services to over 2 million clients via desktop and mobile applications across multiple platforms.",
   "1376955476": "Swap charge",
-  "1380047398": "Scan to download",
   "1382453331": "Payment-related vulnerabilities that could result in loss to our clients and the company",
   "1384941966": "Citigroup",
   "1385437335": "We offer two trading platforms in the EU: <0>Deriv MT5</0> and <2>DTrader</2>. Each of these platforms is designed to fit any trading style, regardless of your trading experience.",
@@ -1583,7 +1582,6 @@
   "2075451186": "If you’re already registered as a Deriv affiliate, you can email us at <0>affiliates@deriv.com</0> instead.",
   "2075608900": "Competitive fixed and variable spreads",
   "2075864188": "FREE",
-  "2076629094": "Android & Huawei",
   "2076930425": "Find out how Deriv P2P works",
   "2077489833": "Do I get commission as a payment agent?",
   "2080629326": "Select your XML file and hit <0>Open</0>.",
