+++ src/pages/deriv-go/_banner.tsx
@@ -1,32 +1,16 @@
 import React from 'react'
 import styled from 'styled-components'
 import DerivGoRow from 'images/common/deriv-go/banner_image.png'
+import DerivGoRowMobile from 'images/common/deriv-go/deriv-go-mobile.png'
 import CommonHeaderSection from 'components/elements/common-header-section'
 import DerivGOLogo from 'images/svg/deriv-go/deriv-go-banner-logo.svg'
 import device from 'themes/device'
 import useBreakpoints from 'components/hooks/use-breakpoints'
 import useHandleSignup from 'components/hooks/use-handle-signup'
 import useAuthCheck from 'components/hooks/use-auth-check'
-import { LinkButton } from 'components/form'
-import { Localize } from 'components/localization'
 import { handleGetTrading } from 'components/layout/nav/util/nav-methods'
-import useRegion from 'components/hooks/use-region'
 import Shape from 'components/custom/_hero-shape'
 import Button from 'components/custom/_button'
-
-type DHeroProps = {
-    background_alt?: string
-    background_image_name?: string
-    background_svg?: string
-    content?: string | JSX.Element
-    is_live_demo?: boolean
-    image_name?: string
-    is_mobile?: boolean | string
-    hide_signup_login?: boolean
-    join_us_for_free?: boolean
-    Logo?: string
-    title?: string | JSX.Element
-}
 
 //TODO: (deriv-rebranding) to make the content section reusable .
 
@@ -80,27 +64,6 @@ const HeroImageWrapper = styled.div`
         width: 100%;
     }
 `
-const GoToLiveDemo = styled(LinkButton)`
-    color: var(--color-red);
-    border-color: var(--color-red);
-    padding: 11px 16px;
-    width: auto;
-    border-radius: 16px;
-
-    &:hover {
-        color: var(--color-white);
-        background-color: var(--color-red);
-    }
-    @media ${device.tablet} {
-        max-width: 100%;
-        white-space: nowrap;
-        margin-left: 0;
-        width: 100%;
-    }
-    @media (max-width: 360px) {
-        white-space: nowrap;
-    }
-`
 const CreateAccountButton = styled(Button)`
     @media ${device.tablet} {
         max-width: 100%;
@@ -119,18 +82,24 @@ const Content = styled.div`
         padding: 0 16px 64px;
     }
 `
+const StyledTradingLogin = styled.img`
+    width: 237px;
+    height: 64px;
 
-const DHero = ({ join_us_for_free, is_live_demo, image_name }: DHeroProps) => {
-    const getLinkType = () => (image_name === 'dbot' ? 'dbot' : 'deriv_app')
+    @media ${device.tablet} {
+        width: 119px;
+        height: 32px;
+    }
+`
+const DHero = () => {
     const { is_mobile } = useBreakpoints()
-    const { is_eu, is_row } = useRegion()
     const handleSignup = useHandleSignup()
     const [is_logged_in] = useAuthCheck()
     return (
         <BackgroundStyle>
             <ContentWrapperStyle>
                 <Content>
-                    <img width="237px" height="64px" src={DerivGOLogo} />
+                    <StyledTradingLogin src={DerivGOLogo} />
                     <CommonHeaderSection
                         title="_t_A trading platform for on-the-go traders_t_"
                         title_font_size={`${is_mobile ? 32 : 64}px`}
@@ -158,7 +127,7 @@ const DHero = ({ join_us_for_free, is_live_demo, image_name }: DHeroProps) => {
             <HeroImageWrapper>
                 <Shape angle={is_mobile ? 101 : 163}>
                     <ImageWrapper>
-                        <ImageStyle src={DerivGoRow} />
+                        <ImageStyle src={is_mobile ? DerivGoRowMobile : DerivGoRow} />
                     </ImageWrapper>
                 </Shape>
             </HeroImageWrapper>
